v0.4: capability release — pillars, bridges, hardening, tooling
v0.4 is a foundation release that transforms the repo from a spin-model
demo into a capability platform. Highlights:

Pillar scaffolds
  src/nqs/          ViT-ready NQS substrate (MLP, real RBM, complex RBM
                    ansätze, Metropolis + QGT stochastic reconfiguration)
  src/mps/          MPS + 2-site DMRG + TEBD imaginary-time + Jacobi SVD
                    + Lanczos; matches ED to ~1e-14 on N≤20 Heisenberg
  src/llg/          LLG time-integrator with exchange + demag fields
  src/equivariant_gnn/
                    torque-network scaffolding over libirrep
  src/neural_operator/
                    2D FNO + FFT + identity / low-pass modes
  src/flow_matching/
                    conditional and unconditional flow-matching sampler
  src/qec_decoder/  greedy-matching QEC baseline
  src/fibonacci_anyons/
                    F- and R-matrix algebra for braid compilation
  src/neuromorphic/ p-bit and Hopfield energy-based solvers
  src/thermodynamic/ RBM-CD + Ising MaxCut annealer
  src/thqcp/        coupled thqcp classical–quantum pillar

Bridges (dormant by default; enabled via build flags)
  engine_adapter, eshkol_bridge, libirrep_bridge, moonlab_bridge,
  noesis_bridge, qgtl_bridge, qllm_bridge

Core library hardening
  Input validation + NULL checks + partial-allocation rollback on every
  initialize_*/create_* constructor (ising, kitaev, spin, neural_network,
  berry_phase, majorana_hilbert_state, toric_code, thqcp, nqs_sampler,
  nqs_ansatz, rbm_cd, fno_layer_2d, hopfield, etc.).
  size_t casts on all multi-dim allocation sizes to eliminate int-
  multiplication overflow at wide layers / large grids.
  Deterministic RNG: --seed flag on the demo binary, SEED env on the
  auxiliary binaries; default to a fixed integer constant. Tests use
  per-test fixed seeds.

Build + tooling
  Makefile: consolidated pillar source-group block (no forward-
  references); SDL resolution via pkg-config with a documented
  fallback; -Wall → -Wall -Wextra baseline with CFLAGS_EXTRA opt-in;
  test_asan target wrapping SANITIZE=1 with UBSAN_OPTIONS presets;
  sdl_check diagnostic target.
  CI (.github/workflows/test.yml): ubuntu + macos build/test matrix
  plus a separate sanitizer-instrumented job.
  REPRODUCIBILITY.md: deterministic-RNG and floating-point discipline.
  DEVELOPMENT.md, SECURITY.md, VERSION_PINS: baseline release docs.

Documentation
  docs/architecture_v0.4.md describes the v0.4 module map and v0.5+
  capability roadmap. docs/toric_code.md, docs/majorana_zero_modes.md,
  docs/berry_phase.md, docs/nqs.md, docs/testing.md were reconciled
  against the shipped API (stale listings, nonexistent functions, and
  out-of-sync struct definitions removed or replaced).

Tests
  One assert-based suite per library module (`tests/test_*.c`), plus
  pillar suites for the scaffolds above. Run via `make test`.

v0.4 is the stable foundation for v0.5's research-pillar work.

+++ src/ising_chain_qubits.c
@@ -25,7 +25,7 @@ IsingChainQubits* initialize_ising_chain_qubits(KitaevLattice *lattice,
     qubits->num_chains = num_chains;
     
     // Allocate memory for chains
-    qubits->chains = (MajoranaChain **)malloc(num_chains * sizeof(MajoranaChain *));
+    qubits->chains = (MajoranaChain **)malloc((size_t)num_chains * sizeof(MajoranaChain *));
     if (!qubits->chains) {
         fprintf(stderr, "Error: Memory allocation failed for chains\n");
         free(qubits);
@@ -33,7 +33,7 @@ IsingChainQubits* initialize_ising_chain_qubits(KitaevLattice *lattice,
     }
     
     // Allocate memory for chain positions
-    qubits->chain_positions = (int (*)[3])malloc(num_chains * sizeof(int[3]));
+    qubits->chain_positions = (int (*)[3])malloc((size_t)num_chains * sizeof(int[3]));
     if (!qubits->chain_positions) {
         fprintf(stderr, "Error: Memory allocation failed for chain_positions\n");
         free(qubits->chains);
@@ -42,7 +42,7 @@ IsingChainQubits* initialize_ising_chain_qubits(KitaevLattice *lattice,
     }
     
     // Allocate memory for chain orientations
-    qubits->chain_orientations = (int *)malloc(num_chains * sizeof(int));
+    qubits->chain_orientations = (int *)malloc((size_t)num_chains * sizeof(int));
     if (!qubits->chain_orientations) {
         fprintf(stderr, "Error: Memory allocation failed for chain_orientations\n");
         free(qubits->chain_positions);
